Update Sysmon Install To Prevent Insecure ACLs
When the .zip file was being copied over to the C:\Program Files directory it was carrying a ACL rule allowing modify access to the folder for the NT AUTHORITY\Authenticated Users group. By doing it this way the only ACLs placed on the C:\Program Files\Sysmon directory will be the ones inherited by C:\Program Files.